refactor(app): Coordinator 구조 개편 — MainCoordinator 도입 및 RecordingCoordinator 제거
- AppCoordinator를 진입 분기 역할로 단순화 (OnboardingCoordinatorDelegate 직접 구현)
- MainCoordinator 도입: 메인 플로우 + RecordingCoordinating 채택으로 녹음 플로우 통합 관리
- RecordingCoordinator 제거: 단일 화면 dismiss 플로우로 1:1 Coordinator 불필요
- presenter를 외부에서 주입받도록 변경 (RecordingCoordinator default parameter 제거)
